# openapi.yaml
openapi: 3.0.1
info:
  title: TravelSite
  description: ""
  termsOfService: ""
  contact:
    email: "hinz@gmail.de"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:3000
tags:
  - name: trips
    description: Everything about your trips

paths:
  /trips:
    post:
      tags:
        - trips
      summary: Add a new trip
      requestBody:
        description: Trip object that needs to be added
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trip"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrip"
    get:
      tags:
        - trips
      summary: Get all trips
      responses:
        "200":
          description: Successfull request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedTrip"

  /trips/{tripId}:
    delete:
      tags:
        - trips
      summary: Delete a trip by ID
      parameters:
        - name: tripId
          in: path
          description: ID of trip to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: successful operation
        "400":
          description: Bad request
          content: {}
        "404":
          description: Trip not found
          content: {}
    patch:
      tags:
        - trips
      summary: update a trip by ID
      parameters:
        - name: tripId
          in: path
          description: ID of trip to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Trip object that needs to be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trip"
        required: true
      responses:
        "204":
          description: successful operation
        "400":
          description: Bad request
          content: {}
        "404":
          description: Trip not found
          content: {}

  /login:
    post:
      tags:
        - login
      summary: Log in and get cookie for authentication
      responses:
        "200":
          description: Successfull request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Login"

  /trips/{userIDtrips}:
    post:
      tags:
        - trips
      summary: Add a new trip to user with userID
      requestBody:
        description: Trip object that needs to be added
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trip"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedTrip"
    get:
      tags:
        - trips
      summary: Get all trips from specific user
      responses:
        "200":
          description: Successfull request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedTrip"

components:
  schemas:
    Trip:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        start:
          type: string
          format: date
        end:
          type: string
          format: date
        country:
          type: string
    SavedTrip:
      allOf:
        - $ref: "#/components/schemas/Trip"
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              format: uuid
    Login:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string